% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaM.RC.out.plot.R
\name{BaM.RC.out.plot}
\alias{BaM.RC.out.plot}
\title{Plot rating curve outputs}
\usage{
BaM.RC.out.plot(DIRPATH = Sys.getenv("DIRPATH"),
  BAMWS = Sys.getenv("BAMWS"), Hgrid)
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in 
the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config 
files in the Docker container [string]}

\item{Hgrid}{A numeric array of the stage values overwhich the rating curve was evaluated 
and will be plotted [numeric]}
}
\value{
This function writes out configurations and runs BaM in prediction mode
}
\description{
This function plots and saves figures from the BaM rating curve prediction 
outputs
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
