% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaM.run.pred.RC.R
\name{BaM.run.pred.RC}
\alias{BaM.run.pred.RC}
\title{Run predictions for rating curve}
\usage{
BaM.run.pred.RC(DIRPATH = Sys.getenv("DIRPATH"),
  BAMFOLD = Sys.getenv("BAMFOLD"), BAMFILE = Sys.getenv("BAMFILE"),
  BAMWS = Sys.getenv("BAMWS"), minH, maxH)
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in 
the Docker container [string]}

\item{BAMFOLD}{An environment variable that contains the folder name of the BaM exe in 
the Docker container [string]}

\item{BAMFILE}{An environment variable that contains the name of the BaM exe in 
the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config 
files in the Docker container [string]}

\item{minH}{Minimum stage value to evaluate rating curve over [numeric]}

\item{maxH}{Maximum stage value to evaluate rating curve over [numeric]}
}
\value{
This function writes out configurations and runs BaM in prediction mode
}
\description{
This function writes out configuration files and runs BaM in prediction 
mode for ratings curves
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
