% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.cont.strm.Q.R
\name{calc.cont.strm.Q}
\alias{calc.cont.strm.Q}
\title{Function to calculate Continuous Discharge}
\usage{
calc.cont.strm.Q(
  DIRPATH = Sys.getenv("DIRPATH"),
  BAMFOLD = Sys.getenv("BAMFOLD"),
  BAMFILE = Sys.getenv("BAMFILE"),
  DATAWS = Sys.getenv("DATAWS"),
  BAMWS = Sys.getenv("BAMWS"),
  startDate = Sys.getenv("STARTDATE"),
  site = Sys.getenv("SITE")
)
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in
the Docker container [string]}

\item{BAMFOLD}{An environment variable that contains the folder name of the BaM exe in
the Docker container [string]}

\item{BAMFILE}{An environment variable that contains the name of the BaM exe in
the Docker container [string]}

\item{DATAWS}{An environment variable that contains the location of the downloaded NEON
zip files in the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config
files in the Docker container [string]}

\item{startDate}{A date in the format of "YYYY-MM-DD", entered as env variable in
run command [string]}

\item{site}{A four letter site code for a NEON site, entered as env variable in
run command [string]}
}
\value{
This function results in text files being written to a directory that contain
the continuous data.
}
\description{
This script uses the BaM executable to predict continuous stage and flow series
with associated uncertainties for a site in a given month using NEON data.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
}
