% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.gaug.press.rel.R
\name{calc.gaug.press.rel}
\alias{calc.gaug.press.rel}
\title{Function to calculate Gauge-Pressure relationship}
\usage{
calc.gaug.press.rel(
  DIRPATH = Sys.getenv("DIRPATH"),
  BAMWS = Sys.getenv("BAMWS"),
  DATAWS = Sys.getenv("DATAWS"),
  startDate = Sys.getenv("STARTDATE"),
  site = Sys.getenv("SITE")
)
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in
the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config
files in the Docker container [string]}

\item{DATAWS}{An environment variable that contains the location of the downloaded NEON
zip files in the Docker container [string]}

\item{startDate}{A date in the format of "YYYY-MM-DD", entered as env variable in
run command [string]}

\item{site}{A four letter site code for a NEON site, entered as env variable in
run command [string]}
}
\value{
This function results in a text file being written to DATAWS.
}
\description{
This script uses multiple data products to establish a relationship between
stream stage and calibrated pressure for a given water year
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Zachary Nickerson \email{nickerson@battelleecology.org} \cr
}
