% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.stag.Q.curv.R
\name{calc.stag.Q.curv}
\alias{calc.stag.Q.curv}
\title{Function to calculate Stage-Discharge Rating Curve}
\usage{
calc.stag.Q.curv(DIRPATH = Sys.getenv("DIRPATH"),
  BAMFOLD = Sys.getenv("BAMFOLD"), BAMFILE = Sys.getenv("BAMFILE"),
  DATAWS = Sys.getenv("DATAWS"), BAMWS = Sys.getenv("BAMWS"),
  startDate = Sys.getenv("STARTDATE"), site = Sys.getenv("SITE"))
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in 
the Docker container [string]}

\item{BAMFOLD}{An environment variable that contains the folder name of the BaM exe in 
the Docker container [string]}

\item{BAMFILE}{An environment variable that contains the name of the BaM exe in 
the Docker container [string]}

\item{DATAWS}{An environment variable that contains the location of the downloaded NEON 
zip files in the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config 
files in the Docker container [string]}

\item{startDate}{A date in the format of "YYYY-MM-DD", entered as env variable in 
run command [string]}

\item{site}{A four letter site code for a NEON site, entered as env variable in 
run command [string]}
}
\value{
This function results in text files being written to a directory that contain 
the posterior parameters and residuals for a stage-discharge rating curve.
}
\description{
This script uses the BaM executable to calculate the parameters that define 
the stage-discharge relationship for a site and water year using NEON data.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
