% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.calc.Q.R
\name{conv.calc.Q}
\alias{conv.calc.Q}
\title{Convert units and re-calculate discharge}
\usage{
conv.calc.Q(stageData, dischargeData)
}
\arguments{
\item{stageData}{A dataframe containing stage and discharge data for a site/water year [dataframe]}

\item{dischargeData}{A dataframe containing distance and velocity point measurements for the 
site/water year that can be used to re-calculate discharge for the stageData [dataframe]}
}
\value{
This function returns an XML file parsed into a data frame
}
\description{
This function GETs data from the NEON REST Service after retrieving a 
transition object that must be set as an environment variable called "TRANSID" using 
two other environment variables that store OS (called osdataserviceurl, required) and 
IS (called cdsurl, optional) REST URLs.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
