% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.WY.strt.end.date.R
\name{def.calc.WY.strt.end.date}
\alias{def.calc.WY.strt.end.date}
\title{Find water year start and end date for an input date}
\usage{
def.calc.WY.strt.end.date(searchIntervalStartDate)
}
\arguments{
\item{searchIntervalStartDate}{A date in POSIXct format [POSIXct]}
}
\value{
List of start and end dates of the water year for the input searchIntervalStartDate
}
\description{
This function returns the start and end dates of the water year for the
input searchIntervalStartDate
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
}
