% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmt.pre.parm.file.R
\name{frmt.pre.parm.file}
\alias{frmt.pre.parm.file}
\title{Format prior paramaters file}
\usage{
frmt.pre.parm.file(dataFrame, numCtrls)
}
\arguments{
\item{dataFrame}{A dataframe containing data to be formatted [dataframe]}

\item{numCtrls}{Number of hydraulic controls [integer]}
}
\value{
This function returns a dataframe formatted identical to the similar publication
table in the Stage-discharge rating curve (DP4.00133.001) data product.
}
\description{
This function takes a dataframe and formats it for the NEON format.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
}
