% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmt.stag.disc.curv.info.file.R
\name{frmt.stag.disc.curv.info.file}
\alias{frmt.stag.disc.curv.info.file}
\title{Format Curve Info File}
\usage{
frmt.stag.disc.curv.info.file(dataFrame, metadata)
}
\arguments{
\item{dataFrame}{A dataframe containing data to be formatted [dataframe]}

\item{metadata}{A list containing the transition metadata that includes: domain,site,
startDateFormatted,endDateFormatted,namedLocationName,numCtrls,numCurves,waterYear [list]}
}
\value{
This function returns a dataframe formatted identical to the similar publication
table in the Stage-discharge rating curve (DP4.00133.001) data product.
}
\description{
This function takes a dataframe and .
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
}
