% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.sim.plot.R
\name{MCMC.sim.plot}
\alias{MCMC.sim.plot}
\title{Plot Marcov Chain Monte Carlo Simulations}
\usage{
MCMC.sim.plot(
  DIRPATH = Sys.getenv("DIRPATH"),
  BAMWS = Sys.getenv("BAMWS"),
  curveID,
  numCtrls,
  Results_MCMC_Cooked,
  NEONformat = F
)
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in
the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config
files in the Docker container [string]}

\item{curveID}{Unique identifier of active stage-discharge rating curve [string]}

\item{numCtrls}{Number of hydraulic controls [integer]}

\item{Results_MCMC_Cooked}{A dataframe containing the posterior MCMC samples,
AKA cooked spaghettis [dataframe]}

\item{NEONformat}{An indicator of the format of the priorParams and Results_MCMC_Cooked
dataframes, for NEON downloads = TRUE, for data firectly form a BaM run = FALSE [boolean]}
}
\value{
This function writes out .PDF files visualizing MCMC simulations
}
\description{
This function plots MCMC simulations from BaM outputs or
NEON data downloads
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
