% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.post.parm.plot.R
\name{pre.post.parm.plot}
\alias{pre.post.parm.plot}
\title{Plot prior and posterior parameter distributions}
\usage{
pre.post.parm.plot(
  DIRPATH = Sys.getenv("DIRPATH"),
  BAMWS = Sys.getenv("BAMWS"),
  curveID,
  numCtrls,
  priorParams,
  Results_MCMC_Cooked,
  NEONformat = F
)
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in
the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config
files in the Docker container [string]}

\item{curveID}{Unique identifier of active stage-discharge rating curve [string]}

\item{numCtrls}{Number of hydraulic controls [integer]}

\item{priorParams}{A dataframe containing the prior parameters [dataframe]}

\item{Results_MCMC_Cooked}{A dataframe containing the posterior MCMC samples,
AKA cooked spaghettis [dataframe]}

\item{NEONformat}{An indicator of the format of the priorParams and Results_MCMC_Cooked
dataframes, for NEON downloads = TRUE, for data firectly form a BaM run = FALSE [boolean]}
}
\value{
This function read in and writes out a modified version of the RunOptions for BaM
}
\description{
This function plot prior and posterior parameters from BaM outputs or
NEON data downloads
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
