% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt.fmt.MCMC.cooked.R
\name{txt.fmt.MCMC.cooked}
\alias{txt.fmt.MCMC.cooked}
\title{Formats the MCMC cooked data to be 15 characters wide}
\usage{
txt.fmt.MCMC.cooked(inputNum)
}
\arguments{
\item{inputNum}{input number in character format [character]}
}
\value{
This writes out number formatted as 15 character scientific notation string
}
\description{
This function returns a number formatted as 15 character scientific notation.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
