% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt.out.ctrl.and.prior.parm.ext.R
\name{txt.out.ctrl.and.prior.parm.ext}
\alias{txt.out.ctrl.and.prior.parm.ext}
\title{Writes out the Config_ControlMatrix and Config_Model text files for BaM}
\usage{
txt.out.ctrl.and.prior.parm.ext(site, searchIntervalStartDate,
  searchIntervalEndDate, controlMatrixPath, priorParamsPath,
  downloadedDataPath)
}
\arguments{
\item{site}{4 letter SITE code [string]}

\item{searchIntervalStartDate}{Start date for search interval [POSIXct]}

\item{searchIntervalEndDate}{Start date for search interval [POSIXct]}

\item{controlMatrixPath}{File path to write out controlMatrix file [string]}

\item{priorParamsPath}{File path to write out Model file [string]}

\item{downloadedDataPath}{File path where downloaded .zip files can be found [string]}
}
\value{
This function read in and writes out a modified version of the RunOptions text 
file for BaM and returns the number of controls and inputResultUuids for the control 
matrix and prior parameters tables
}
\description{
This function writes out the ControlMatrix and Model configurations for BaM.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
