% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt.out.run.opts.R
\name{txt.out.run.opts}
\alias{txt.out.run.opts}
\title{Writes out a text file with RunOptions for BaM}
\usage{
txt.out.run.opts(runType, RunOptionsPath)
}
\arguments{
\item{runType}{Either "param" or "pred" to determine which RunOptions to enable [string]}

\item{RunOptionsPath}{A filepath for the RunOptions text file [string]}
}
\value{
This function read in and writes out a modified version of the RunOptions for BaM
}
\description{
This function writes out the RunOptions for BaM.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
