% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt.out.spag.data.R
\name{txt.out.spag.data}
\alias{txt.out.spag.data}
\title{Writes out a text file with "cooked spaghettis" for BaM}
\usage{
txt.out.spag.data(spagDataIn, spagOutPath)
}
\arguments{
\item{spagDataIn}{Spaghettis, "sampled parameters", for the rating curve used to predict Q [dataFrame]}

\item{spagOutPath}{Filepath to write re-formatted file out to including the filename [string]}
}
\value{
This writes out txt file of "cooked spaghettis" for BaM
}
\description{
This function writes out the "cooked spaghettis" for BaM.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
