% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmt.pre.parm.file.R
\name{frmt.pre.parm.file}
\alias{frmt.pre.parm.file}
\title{Format MCMC cooked results File}
\usage{
frmt.pre.parm.file(dataFrame, numCtrls)
}
\arguments{
\item{dataFrame}{A dataframe containing data to be formatted [dataframe]}

\item{numCtrls}{Number of hydraulic controls [integer]}
}
\value{
This function returns a dataframe formatted for integration into the transition object
}
\description{
This function takes a dataframe and .
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
