% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.sim.plot.R
\name{MCMC.sim.plot}
\alias{MCMC.sim.plot}
\title{Plot prior and posterior parameter distributions}
\usage{
MCMC.sim.plot(DIRPATH = Sys.getenv("DIRPATH"),
  BAMWS = Sys.getenv("BAMWS"), numCtrls, Results_MCMC_Cooked, NEONformat)
}
\arguments{
\item{DIRPATH}{An environment variable that contains the location of the files in 
the Docker container [string]}

\item{BAMWS}{An environment variable that contains the location of the BaM config 
files in the Docker container [string]}

\item{numCtrls}{Number of hydraulic controls [integer]}

\item{Results_MCMC_Cooked}{A dataframe containing the posterior MCMC samples, 
AKA cooked spaghettis [dataframe]}

\item{NEONformat}{An indicator of the format of the priorParams and Results_MCMC_Cooked 
dataframes, for NEON downloads = TRUE, for data firectly form a BaM run = FALSE [boolean]}
}
\value{
This function read in and writes out a modified version of the RunOptions for BaM
}
\description{
This function plot prior and posterior parameters from BaM outputs or 
NEON data downloads
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
